Ensure that playPause toggles playback in any state
It can be assumed that an user wants to start playback when pressing the
play button. However, just toggling the pause state is no good if the
current mpd state is stopped or unknown. When this situation happens,
playback is now started in position 0 of the queue to ensure something
starts playing.