Send activation notification on click events
MediaKeyTap listens to NSWorkspaceDidActivateApplication notifications
to determine which application is the most recently activated
application and thus needs to be listened to for media key events. Now
kmbmpdc refreshes its position in the MediaKeyTap application stack
every time one of the three UI elements on the menu bar are clicked.

A changelog file (CHANGELOG.md) is also added based on the "Keep a
CHANGELOG" (http://keepachangelog.com/) format.