Fixes to track change notifications
AppDelegate is set as the NSUserNotificationCenter delegate and always
returns `true` in the `userNotificationCenter` method in order to force
Notification Center always show notifications for kmbmpdc. In addition
all previous notifications are removed upon application launch.

When mpd playback is stopped or unknown, the current track is set to
`nil`. This is to ensure that a track notification is presented when
playback is restarted, possibly long after the track last changed.