Switch to a forked version of MediaKeyTap
The previous solution of sending out activation notifications was not
really a functional one and other applications might steal focus away
from kmbmpdc without of it ever regaining it. In order to fix this, the
application is made to use a custom forked version of MediaKeyTap that
can manually set itself as the top application on the listener stack.