Add automatic reconnection
When the client is disconnected without user interaction, the client
starts automatically attempting a new connection to the server. The
client waits 2 seconds before the first attempt and then doubles the
wait time for subsequent attempts with a maximum wait time of 60
seconds. When a connection is established, the timer is reset.