Add upcoming track list to controller
The upcoming track list is implemented as a NSTableView, which gets its
data from a singleton TrackQueue object. The track data is presented as
TrackCell objects, which subclass from NSTableCellView. Track class is
edited so that it can function for both current track and queue tracks.

MPDController now watches for MPD_IDLE_QUEUE events in the idle loop to
know when the queue is edited. New queue data is fetched by
MPDController.reloadQueue(). Whenever the queue is updated by the
MPDController, KMBMPDCQueueReload notification is sent out so that the
user interface knows to reload the table data.

Idle loop logic is improved to catch multiple idle events during one
loop. This is because during consume mode, both MPD_IDLE_QUEUE and
MPD_IDLE_PLAYER events get sent at the same time.