Change queue fetching to prevent race conditions
If two queue reloads happen nearly at the same time, the NSTableView
might be updating just as the queue is getting wiped, resulting in an
index out of range exception as the table view tries to fetch now
removed items from the TrackQueue. This change should avoid overlapping
reloads, emptying the track queue for longer than needed and generating
table cells for indexes outside the queue.