Simplify track queue design and restore queue UI
Cover art is removed from the track queue cell design to have a less of
an impact on memory and disk usage. Not having to load the cover art
dynamically as the list is being scrolled also improves scrolling
performance for the list.

Additionally, the open/closed state of the track queue is stored in
settings and restored when the application is loaded.