Fit cover art vertically in controller
Cover art view is now resized vertically so that it fits the cover art
aspect ratio, with a minimum height of 89 points and a maximum height of
600 points.

Cover art images are also scaled with the highest quality interpolation.
This is especially helpful for large scanned cover art, which can look
atrocious when directly displayed in the controller.