Produce two image representations for cover art
In order to display good quality cover art on both on Retina and
non-Retina displays, two different sized images are needed. The old code
used to only produce one size image that is 300 points wide, which
translates to 300 pixels on non-Retina displays and 600 pixels on Retina
displays. When using a Retina display in conjuction with a non-Retina
display, the user would use the 600 pixel version on both displays,
producing another (low-quality) downscale from 600 pixels to 300 pixels.

This change still produces one NSImage to use with the NSImageView, but
produces the images by creating bitmap representations at 300 pixels and
600 pixels respectively. NSImageView can then choose the corresponding
image depending on the display's resolution.